backend/requirements.txt
fastapi
uvicorn
sqlalchemy
groq
pydantic
python-dotenv
